+++ vertx-server/src/main/scala/com/cloudentity/tools/vertx/server/api/ApiServerRoot.scala
@@ -9,7 +9,7 @@ import com.cloudentity.tools.vertx.scala.bus.ScalaComponentVerticle
 import com.cloudentity.tools.vertx.server.api.conf.{ApiServerConf, FilterConf}
 import com.cloudentity.tools.vertx.server.api.conf.codecs._
 import com.cloudentity.tools.vertx.server.api.filters.RouteFilter
-import io.vertx.core.{DeploymentOptions, Vertx}
+import io.vertx.core.Vertx
 import com.cloudentity.tools.vertx.verticles.VertxDeploy
 import org.slf4j.LoggerFactory
 
@@ -18,7 +18,7 @@ import scala.concurrent.{ExecutionContext, Future}
 class ApiServerRoot() extends ScalaComponentVerticle {
   val log = LoggerFactory.getLogger(this.getClass)
 
-  override def verticleId(): String = Option(super.verticleId()).getOrElse("apiServer")
+  override def verticleId(): String = Option(super.verticleId()).getOrElse(ApiServer.defaultVerticleId)
 
   override def initComponentAsync(): VxFuture[Void] = {
       decode[ApiServerConf](getConfig().toString).left.map(_.fillInStackTrace()) match {
@@ -29,8 +29,8 @@ class ApiServerRoot() extends ScalaComponentVerticle {
               _ <- RegistryVerticle.deploy(vertx, serverConf.routesRegistry.getOrElse("routes")).toScala()
               _ <- RegistryVerticle.deploy(vertx, serverConf.filtersRegistry.getOrElse("filters"), false).toScala()
               _ <- validateFilterConfs(vertx, serverConf.routes.value.flatMap(_.filters.getOrElse(Nil)))
-              _ <- deployServers(vertx, serverConf)
-            } yield (())
+              _ <- deployServer(vertx, serverConf)
+            } yield ()
         case Left(ex) =>
           log.error(s"'${verticleId()}' configuration is wrong. Decoding ApiServerConf failed", ex)
           Future.failed(new Exception(s"'${verticleId()}' configuration is wrong", ex))
@@ -51,23 +51,7 @@ class ApiServerRoot() extends ScalaComponentVerticle {
       }
     }
 
-  /**
-    * Deploys DeploymentOptions.instances of {@link com.cloudentity.tools.vertx.server.ApiServer} along with the route-handler verticles.
-    *
-    * @param vertx
-    * @param serverConf
-    * @param ec
-    * @return instance of io.vertx.core.Future, will be failed if operation failed, successful otherwise
-    */
-  def deployServers(vertx: Vertx, serverConf: ApiServerConf)(implicit ec: ExecutionContext): Future[Void] = {
-    def deployRec(instances: Int, acc: Future[Unit]): Future[Unit] =
-      if (instances > 0) {
-        val server  = new ApiServer(serverConf)
-        val nextAcc = acc.flatMap(_ => VertxDeploy.deploy(vertx, server, new DeploymentOptions()).toScala.map(_ => ()))
-
-        deployRec(instances - 1, nextAcc)
-      } else acc
-
-    deployRec(serverConf.serverInstances.getOrElse(2 * Runtime.getRuntime().availableProcessors()), Future.successful(())).map(_ => null: Void)
+  def deployServer(vertx: Vertx, serverConf: ApiServerConf)(implicit ec: ExecutionContext): Future[Void] = {
+    VertxDeploy.deploy(vertx, new ApiServer(serverConf), verticleId()).toScala.map(_ => null: Void)
   }
 }