+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [1.6.0] - 2021-02-26
+### Added
+- ConsulSdRegistrar supports dynamic port
+- ConsulSdRegistrar module deploys 'system-ready' registry
+- Ability to override name of the service config for vertx service tests
+
+### Changed
+- Deploying 1 instance of ApiServer instead of 2*CPUs (vertx-server) - improves performance + allows using dynamic port
+- VertxBootstrap starts 'system-init' registry before server start and 'system-ready' after server start (vertx-server)
+- SmartHttpClientImpl body stream resetting order changed in case of exception
+
 ## [1.5.0] - 2021-02-04
 ### Added
 - GetConfigRoute implementation returning masked global config